Persistência e testes de integração

+++ kafka-producer/src/main/java/com/github/ricardocomar/kafkabalancedconsumers/kafkaproducer/entrypoint/model/ProcessResponse.java
@@ -9,5 +9,5 @@ public class ProcessResponse {
 
 	private String id;
 	private String responseId;
-	private Long duration;
+	private Integer duration;
 }
